Move back to index.ts for a cleaner diff
